add revert form function, move delete btn
