+++ client/src/components/forms/UpdateForm.jsx
@@ -5,7 +5,7 @@ import { UPDATE_PRODUCT, DELETE_PRODUCT } from "../../utils/mutations";
 
 function UpdateForm(prop) {
   const [UpdateProduct] = useMutation(UPDATE_PRODUCT);
-  const [DeleteProduct] = useMutation(DELETE_PRODUCT);
+  // const [DeleteProduct] = useMutation(DELETE_PRODUCT);
   const { description, category, image, name, price, quantity } = prop.product;
   const { productId } = useParams();
 
@@ -45,8 +45,7 @@ function UpdateForm(prop) {
   //TODO FINISH FUNCTION
   const handleFormSubmit = async (event) => {
     event.preventDefault();
-    // console.log(formState);
-    // const { name, price, description, image, category, quantity } = formState;
+
     try {
       const { data } = await UpdateProduct({
         variables: formState,
@@ -55,23 +54,24 @@ function UpdateForm(prop) {
       console.log(err);
     }
   };
-  const deleteItem = async (event) => {
-    event.preventDefault;
-    try {
-      const { data } = await DeleteProduct({
-        variables: { id: productId },
-      });
-      if (data.deleteProduct) {
-        window.location.assign("/");
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
+
+  function revertEdit() {
+    setFormState({
+      id: productId,
+      name: name,
+      price: price,
+      category: category,
+      description: description,
+      quantity: quantity,
+      image: image,
+    });
+  }
+
   return (
     <>
       <form onSubmit={handleFormSubmit}>
         <h2>Edit product</h2>
+        
         <label htmlFor="productId">Product ID:(editing disabled)</label>
         <input
           value={formState._id}
@@ -130,17 +130,18 @@ function UpdateForm(prop) {
           disabled
         ></input>
 
+
+      </form>
+      <div className="form-footer center">
         <div className="button-container">
-          <button className="btn-2" onClick={deleteItem}>
-            Delete Product
+          <button className="btn-2" onClick={revertEdit}>
+            Revert
           </button>
           <button className="btn-1" type="submit">
             Save Changes
           </button>
         </div>
-
-        <div className="form-footer"></div>
-      </form>
+      </div>
     </>
   );
 }
